Migrate deprecations in core, compiler, idea, tests
Replacing deprecated Char.toInt() with Char.code and
Number.toChar() with Number.toInt().toChar(), where Number is not Int.

KT-23451

+++ compiler/fir/fir-deserialization/src/org/jetbrains/kotlin/fir/deserialization/AbstractAnnotationDeserializer.kt
@@ -243,7 +243,7 @@ abstract class AbstractAnnotationDeserializer(
                 const(kind, value.intValue)
             }
 
-            CHAR -> const(ConstantValueKind.Char, value.intValue.toChar())
+            CHAR -> const(ConstantValueKind.Char, value.intValue.toInt().toChar())
             FLOAT -> const(ConstantValueKind.Float, value.floatValue)
             DOUBLE -> const(ConstantValueKind.Double, value.doubleValue)
             BOOLEAN -> const(ConstantValueKind.Boolean, (value.intValue != 0L))
